+++ ks_includes/KlippyGtk.py
@@ -165,10 +165,12 @@ class KlippyGtk:
         return b
 
     @staticmethod
-    def HomogeneousGrid():
+    def HomogeneousGrid(width=None, height=None):
         g = Gtk.Grid()
         g.set_row_homogeneous(True)
         g.set_column_homogeneous(True)
+        if width != None and height != None:
+            g.set_size_request(width, height)
         return g
 
     @staticmethod