+++ panels/zcalibrate.py
@@ -19,18 +19,21 @@ class ZCalibratePanel(ScreenPanel):
     distance = 1
     distances = ['.01','.05','.1','.5','1','5']
 
+    def __init__(self, screen, title, back=True):
+        super().__init__(screen, title, False)
+
     def initialize(self, panel_name):
         _ = self.lang.gettext
         grid = KlippyGtk.HomogeneousGrid()
 
-        label = Gtk.Label(_("Z Offset") + ": ")
-        label.get_style_context().add_class('temperature_entry')
-        self.labels['zpos'] = Gtk.Label(_("Homing"))
-        self.labels['zpos'].get_style_context().add_class('temperature_entry')
-        box = Gtk.Box()
+        label = Gtk.Label(_("Z Offset") + ": \n")
+        self.labels['zposition'] = Gtk.Label(_("Homing"))
+        box = Gtk.VBox()
+        box.set_vexpand(False)
+        box.set_valign(Gtk.Align.CENTER)
 
         box.add(label)
-        box.add(self.labels['zpos'])
+        box.add(self.labels['zposition'])
 
         zpos = KlippyGtk.ButtonImage('z-offset-decrease',_("Raise Nozzle"))
         zpos.connect("clicked", self.move, "+")
@@ -57,12 +60,10 @@ class ZCalibratePanel(ScreenPanel):
         self.labels["1"].set_active(True)
 
         space_grid = KlippyGtk.HomogeneousGrid()
+        space_grid.set_row_homogeneous(False)
         space_grid.attach(Gtk.Label(_("Distance (mm)") + ":"),0,0,1,1)
         space_grid.attach(distgrid,0,1,1,1)
         space_grid.attach(Gtk.Label(" "),0,2,1,1)
-
-        estop = KlippyGtk.ButtonImage("decrease",_("Emergency Stop"),"color4")
-        estop.connect("clicked", self.emergency_stop)
 
         complete = KlippyGtk.ButtonImage('complete',_('Accept'),'color2')
         complete.connect("clicked", self.accept)
@@ -71,22 +72,23 @@ class ZCalibratePanel(ScreenPanel):
         b.connect("clicked", self.abort)
 
 
-        grid.attach(zpos, 1, 0, 1, 1)
-        grid.attach(box, 0, 1, 2, 1)
-        grid.attach(zneg, 1, 1, 1, 1)
-        grid.attach(estop, 3, 0, 1, 1)
-        grid.attach(complete, 3, 1, 1, 1)
+        #grid.set_row_homogeneous(False)
+        grid.attach(zpos, 0, 0, 1, 1)
+        grid.attach(box, 1, 0, 2, 2)
+        grid.attach(zneg, 0, 1, 1, 1)
+        grid.attach(complete, 3, 0, 1, 1)
         grid.attach(space_grid, 0, 2, 3, 1)
         grid.attach(b, 3, 2, 1, 1)
 
 
-        self.panel = grid
+        self.content.add(grid)
         self._screen.add_subscription(panel_name)
 
     def activate(self):
-        if self._screen.printer.get_stat("toolhead","homed_axes") != "xyz":
-            self._screen._ws.klippy.gcode_script(KlippyGcodes.HOME)
-        self._screen._ws.klippy.gcode_script(KlippyGcodes.PROBE_CALIBRATE)
+        #if self._screen.printer.get_stat("toolhead","homed_axes") != "xyz":
+        #    self._screen._ws.klippy.gcode_script(KlippyGcodes.HOME)
+        #self._screen._ws.klippy.gcode_script(KlippyGcodes.PROBE_CALIBRATE)
+        print("nothing")
 
     def process_update(self, action, data):
         if action != "notify_status_update":
@@ -96,7 +98,7 @@ class ZCalibratePanel(ScreenPanel):
             self.updatePosition(data['toolhead']['position'])
 
     def updatePosition(self, position):
-        self.labels['zpos'].set_text(str(round(position[2],2)))
+        self.labels['zposition'].set_text(str(round(position[2],2)))
 
     def change_distance(self, widget, distance):
         if self.distance == distance:
@@ -121,9 +123,9 @@ class ZCalibratePanel(ScreenPanel):
     def abort(self, widget):
         logger.info("Aborting Z calibrate")
         self._screen._ws.klippy.gcode_script(KlippyGcodes.PROBE_ABORT)
-        self._screen._menu_go_back(widget)
+        self.menu_return(widget)
 
     def accept(self, widget):
         logger.info("Accepting Z calibrate")
         self._screen._ws.klippy.gcode_script(KlippyGcodes.PROBE_ACCEPT)
-        self._screen._menu_go_back(widget)
+        self.menu_return(widget)